requirements.txt
pandas~=2.1
pydantic~=2.5
numpy~=1.26
python-dateutil~=2.8
requests~=2.3